Adds support for pipeline virtualization
Adds main, index.js script to function as a feeder for the enrichment. The
config and data file are there to support the feeder script.
Modified README according to the new changes.
Slimmed down enrichment.js a little to make it easier to read.
Added new used node modules to package file.